WIP: Removed commented code and minor spec file changes
